saga_log/requirements.txt
pulsar-client==3.0.0
protobuf==4.22.0
Faker==17.6.0
psycopg2-binary==2.9.5
sqlalchemy-orm==1.2.4
